+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 3.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Cleanup all possible HTML in user formitted data and convert it to plain text.
+  [cekk]
 
 
 3.0.1 (2024-04-18)